bossrespawner

+++ src/main/java/com/github/L_Ender/cataclysm/blockentities/Cursed_tombstone_Entity.java
@@ -23,6 +23,7 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.gameevent.GameEvent;
 import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.Vec3;
 
@@ -47,6 +48,7 @@ public class Cursed_tombstone_Entity extends BlockEntity {
                 }else{
                     if (!level.isClientSide) {
                         level.setBlock(pos, blockState.setValue(Cursed_Tombstone_Block.POWERED, Boolean.valueOf(true)), 2);
+                        level.gameEvent(GameEvent.BLOCK_CHANGE, pos, GameEvent.Context.of(null, blockState));
                     }
                 }
             }else{
@@ -107,8 +109,9 @@ public class Cursed_tombstone_Entity extends BlockEntity {
                                         }
                                     }
                                 }
-                                level.addFreshEntity(maledictus);
-                                level.destroyBlock(pos, false);
+                                if(level.addFreshEntity(maledictus)){
+                                    level.destroyBlock(pos, false);
+                                }
                             }
                         }
                     }
